Работа над v1.0.4 На странице админа, вкладка Тренировки отремонтировал подсчет участников, Поправил в боте подсчет участников тренировки при отмене. УБРАЛ ОШИБКУ ЗАПУСКА ДУХ ТОКЕНОВ, ТЕПЕРЬ В ТЕРМИНАЛЕ НЕТ ОШИБОК
